+++ src/mo_config/args.ml
@@ -1,5 +1,6 @@
 (* This module contains some argument parsing that is common between
 multiple executables *)
+open Exit
 
 (** suppress documentation *)
 let _UNDOCUMENTED_ doc = "" (* TODO: enable with developer env var? *)
@@ -12,7 +13,7 @@ let string_map flag r desc =
     Arg.String (fun value ->
       let key = !key_ref in
       if Flags.M.mem key !r
-      then (Printf.eprintf "duplicate %s %s" flag key ; exit 1)
+      then fail "duplicate %s %s" flag key
       else r := Flags.M.add key value !r
     )
   ],