+++ src/mo_config/mo_config.ml
@@ -1,0 +1,6 @@
+module Flags = Flags
+module Args = Args
+module Exit = Exit
+module Internal_error = Internal_error
+
+include Exit